extract syncUiToUrl helper from main layout
